# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StudentCourseAPI",
    "version": "1.0.0",
    "description": "API pédagogique pour gérer étudiants et cours (no DB)"
  },
  "tags": [
    {
      "name": "Courses",
      "description": "Gestion des cours"
    }
  ],
  "paths": {
    "/courses": {
      "get": {
        "tags": ["Courses"],
        "summary": "Liste des cours",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Créer un cours",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "teacher": {
                    "type": "string"
                  }
                },
                "required": ["title", "teacher"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Créé"
          },
          "400": {
            "description": "Paramètres invalides"
          }
        }
      }
    },
    "/courses/{id}": {
      "get": {
        "summary": "Récupérer un cours",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Non trouvé"
          }
        }
      },
      "delete": {
        "summary": "Supprimer un cours",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Supprimé"
          },
          "404": {
            "description": "Non trouvé"
          }
        }
      }
    }
  }
}